Separate translation units for CRC and qCheck
CRC and qCheck are different compilation targets now. CRC32 is still
header-only at the moment but this sets it up to have a more private
implementation
