+++ CMakeLists.txt
@@ -119,8 +119,22 @@ elseif( CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang" )
 	)
 endif()
 
+
+add_library(
+	CRC
+	source/CRC/CRC32.cpp
+)
+target_include_directories(
+	CRC
+	INTERFACE
+		include
+	PRIVATE
+		include/CRC
+)
+
 add_executable(
 	qCheck
+	source/qCheck.cpp
 	source/main.cpp
 )
 
@@ -131,6 +145,7 @@ find_package(Threads REQUIRED)
 target_link_libraries(
 	qCheck
 	PRIVATE
+	CRC
 	Threads::Threads
 )
 target_include_directories(
